+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Changed
 * Accordion: Optionele Badge in handle van Accordion Section ([#3936](https://github.com/dso-toolkit/dso-toolkit/issues/3936))
 * Header: Help-knop in menu bij mobiele viewport ([#3017](https://github.com/dso-toolkit/dso-toolkit/issues/3017))
+* **BREAKING** Document Header: vervang icon-only Button door Icon Button ([#3568](https://github.com/dso-toolkit/dso-toolkit/issues/3568))
 
 ## 🧧 Release 100.3.1 - 2026-09-08